+++ grain_delay/src/lowpass.rs
@@ -2,12 +2,12 @@ use super::clip::Clip;
 use std::f32;
 
 pub struct Lowpass {
-  sample_rate: f64,
+  sample_rate: f32,
   z: f32,
 }
 
 impl Lowpass {
-  pub fn new(sample_rate: f64) -> Self {
+  pub fn new(sample_rate: f32) -> Self {
     Self { sample_rate, z: 0. }
   }
   fn mix(&mut self, a: f32, b: f32, interp: f32) -> f32 {
